optimize item meta writes
avoid writing to stack meta when nothing changed.
In an mk3 powerbank there can be an item with low max-charge
and items with much more max-charge. There is no reason for
the meta of the already charged items to be written over
and over again only because there are other items still
charging.